rename strandtest.py to effects.py
