+++ effects.py
